проведение по регистру дат

+++ server/needs/listener.js
@@ -8,12 +8,17 @@ module.exports = function listener($p, log, glob) {
     glob.client = client;
     subscription.listeners.push(async function reflect({db, results, docs, branch, abonent, year}) {
       for(const {doc, prod} of docs) {
-        // при любом изменении документа, удаляем старые записи
-        await client.query(`DELETE FROM areg_needs where register = $1 and register_type = 'doc.calc_order'`, [doc.ref]);
-        if(doc.posted) {
-          // если документ проведён, добавляем новые
-          const sql = await nsql(doc, $p);
-          await client.query(sql);
+        if(doc.class_name === 'doc.calc_order') {
+          // при любом изменении документа, удаляем старые записи
+          await client.query(`DELETE FROM areg_needs where register = $1 and register_type = 'doc.calc_order'`, [doc.ref]);
+          if(doc.posted) {
+            // если документ проведён, добавляем новые
+            const sql = await nsql(doc, $p);
+            await client.query(sql);
+          }
+        }
+        else if(doc.class_name === 'doc.purchase_order') {
+
         }
       }
     });
